refactor(dashboard): move color interpolation related types to /types/plugin
